Fixed the speed thats being changed by the input field. The intervals are now correctly deleted and added.
